… lru_test.go
… simplelru/lru.go
… simplelru/lru_test.go