@kasegao 2022-10-28
Thank you for your great PR.

In order to keep backwward compability, however, wouldn't it be better to keep the existing implementation as suggested by AskAlexSharov's  comment?

https://github.com/hashicorp/golang-lru/pull/101#issuecomment-1269275609