@jefferai 2022-10-28
@kasegao If this is marked as v2 of the lib we don't need backwards compat.