@kasegao 2022-10-28
@jefferai I misunderstood. Thank you!