@jefferai 2022-10-29
I just bumped the workflows version in CI in `main`, you may want to try updating from there and seeing if that helps.