@SoMuchForSubtlety 2022-10-30
> I just bumped the workflows version in CI in `main`, you may want to try updating from there and seeing if that helps.

Rebasing on master gets rid of the `math/rand` warnings, but I still get the errors from `golint`. It's deprecated[^1] and does not have support for generics, leading to it getting confused by generic method signatures.

[^1]: https://github.com/golang/go/issues/38968