@jefferai 2022-10-31
I removed some deprecated linters from the config -- didn't even realize we had a config and weren't using defaults. Hopefully that fixes it.