@SoMuchForSubtlety 2022-10-31
> I removed some deprecated linters from the config -- didn't even realize we had a config and weren't using defaults. Hopefully that fixes it.

Thanks, I rebased on master and at least locally golangci-lint doesn't complain any more.

Is there anything else open I need to change?