@jefferai 2022-11-03
I will do another pass -- as for stuff outstanding I still have a slight preference not to do bare `return` when nil/default values are being returned from Get functions considering we're not actually populating the named return parameters anywhere. Keeping them named is nice from a doc perspective, but we could explicitly return the false/nil values. But it's more a style thing than anything else. 🤷