@SoMuchForSubtlety 2022-11-03
> Keeping them named is nice from a doc perspective, but we could explicitly return the false/nil values.

Do you mean
```go
// Peek returns the key value (or undefined if not found) without updating
// the "recently used"-ness of the key.
func (c *LRU[K, V]) Peek(key K) (value V, ok bool) {
	var ent *entry[K, V]
	if ent, ok = c.items[key]; ok {
		return ent.value, true
	}
	return value, false
}
```
or 
```go
// Peek returns the key value (or undefined if not found) without updating
// the "recently used"-ness of the key.
func (c *LRU[K, V]) Peek(key K) (value V, ok bool) {
	var ent *entry[K, V]
	if ent, ok = c.items[key]; ok {
		return ent.value, true
	}
	var empty V
	return empty, false
}
```

They are the same, but the seconds one makes it clearer that we are returning an empty value (at the cost of an added line).