@jefferai 2022-11-04
Hm. I keep forgetting that V is not necessarily `nil` anymore :-) 

Leave it be.