@jefferai 2022-11-04
Looks great! I have a question for anyone following this:

I'm tempted to cut a final `v0` tag, set this as `v2`, and have `v1` point to a branch that has a README that basically says, use v0 if you need backwards compat or use v2. This would break users that auto-upgrade until they select a specific v0 (or v2) tag, but also pushes people towards using v2 given that it's faster and more capable.

Thoughts?