@bep 2022-11-09
>Thoughts?

I agree (I do, though, wish Go Modules had a way to handle this cleaner).