@jefferai 2022-11-11
@SoMuchForSubtlety Any thoughts/preferences on my above question?